api/requirements.txt
fastapi==0.85.0
uvicorn==0.18.3
SQLAlchemy==1.4.41
psycopg2-binary